apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "awaaz-api"
version = "0.1.0"
description = "Awaaz control plane — FastAPI + Postgres + WhatsApp FSM"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Gelecek Solution", email = "engineering@awaaz.pk" }]
dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "starlette>=0.40",
    "python-multipart>=0.0.12",
    "slowapi>=0.1.9",
    "httpx>=0.27",
    # Validation & settings
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    # DB
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "alembic>=1.13",
    "pgqueuer>=0.10",
    # Cache / queue
    "redis>=5.0",
    # Storage
    "boto3>=1.35",
    "botocore>=1.35",
    # AI / NLP
    "anthropic>=0.40",
    "openai>=1.50",
    "tiktoken>=0.7",
    # Telephony / channels
    "twilio>=9.3",
    # Pakistan-specific
    "phonenumbers>=8.13",
    "dateparser>=1.2",
    # Logging / observability
    "structlog>=24.4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27",
    "prometheus-client>=0.21",
    "sentry-sdk[fastapi]>=2.13",
    # Billing
    "stripe>=10.10",
    # Webhook signing helpers
    "cryptography>=43.0",
    # Misc
    "tenacity>=9.0",
    "orjson>=3.10",
    "typing-extensions>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pytest-httpx>=0.32",
    "respx>=0.21",
    "freezegun>=1.5",
    "ruff>=0.6.9",
    "black>=24.10",
    "mypy>=1.11",
    "types-redis>=4.6",
    "types-requests>=2.32",
    "types-python-dateutil>=2.9",
    "ipython>=8.28",
    "rich>=13.9",
]

[project.scripts]
awaaz-api = "awaaz_api.main:cli"

[tool.setuptools]
packages = ["awaaz_api"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["awaaz_api"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # comprehensions
    "UP",   # pyupgrade
    "ASYNC",
    "S",    # bandit
    "BLE",  # blind-except
    "RUF",
    "TID",
]
ignore = [
    "S101", # asserts ok in tests
    "B008", # function call in default arg — fine for FastAPI Depends
]

[tool.ruff.lint.per-file-ignores]
"awaaz_api/tests/**" = ["S", "B"]
"awaaz_api/migrations/**" = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_untyped_decorators = false
plugins = ["pydantic.mypy"]
mypy_path = "awaaz_api"
exclude = [
    "awaaz_api/migrations/versions/.*",
    "awaaz_api/tests/.*",
]

[[tool.mypy.overrides]]
module = [
    "phonenumbers.*",
    "dateparser.*",
    "twilio.*",
    "pgqueuer.*",
    "stripe.*",
    "tiktoken.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["awaaz_api/tests"]
markers = [
    "integration: requires real provider credentials",
    "eval: end-to-end LLM evaluation suite (slow)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["awaaz_api"]
omit = [
    "awaaz_api/migrations/*",
    "awaaz_api/tests/*",
    "awaaz_api/scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
]
